--- src/pages/general/Planes.tsx
@@ -1,10 +1,0 @@
-
-const Planes:React.FC = () => {
-  return (
-    <>
-      
-    </>
-  )
-}
-
-export default Planes
